Create EPQ Checkers 1.3
Haven't uploaded in a while but there are a large amount of changes. I thought I would have the next stage of the code done sooner but really wanted to finish it before uploading. Still not done with the current goals (Secondary Takes and Full iteration) but very close. Should hopefully be finished with the basic requirements of the project soon