Create 1.3.1
Finally, a running version of the MiniMax Algorithm that can finish the whole recursion at varying depths. I am yet to make a system to evaluate each game state but I am working on that now. I plan on using Alpha-beta pruning to improve performance but for now this works fine.