Create Login - V04
Fixed bug where the insertion sort wasn't copying all values, I made it so if a value is in the correct position it is added to the sorted list (this did not previously occur). The file sort now functions so long as the final value has an enter value (\n). I need to decide whether to fix this rare error or stop it from occurring by not editing the text file apart from through the new Log in system.

I also started to develop the GUI side of the log in system, it now has a main box with log in text and a text box that can enter the "a" letter, need to add more soon - was more a test of functionality. I have capped the username length at 11 characters as that fits perfectly in the box. In the future I need to add the capability to enter all alphabet and number keys. Then a binary search for the username to make sure it doesn't already match another username on file. Then add that user with default stats (0,0,0).